Prevent writing to application-supplied buffers when encoding strings (#485)
* fix generated AsCopy encoder methods for bytes/string fields when offset is non-zero

After calling copyInto(), the offset is always 0.

* prevent generated encoders from copying into application-supplied buffers

The generated encoders contain different ways of setting string-like
fields. Some methods take buffers owned by the application, e.g. passing
in an AsciiSequenceView. Other methods take data that is copied into the
internal buffer, e.g. passing in a CharSequence.

If an application uses both styles, it could first pass in
an AsciiSequenceView (which would point to some kind of input buffer),
then on a later code path, pass in a CharSequence. This will result
in the encoder copying the CharSequence into the input buffer,
corrupting that buffer.

This change takes a reference to the last buffer created and owned by
the encoder for the UnsafeBuffer. On reset, it resets the UnsafeBuffer
to point it back to that buffer.